Improvements: Extract Security Api (#2983)
