+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 1.4.0 (Not released yet)
 ------------------------
+* [#2983](https://github.com/TouK/nussknacker/pull/2983) Extract Permission to extensions-api
 * [#2996](https://github.com/TouK/nussknacker/pull/2996) Flink job registration optimization - do not prepare interpretation function for ending node
 * [#3010](https://github.com/TouK/nussknacker/pull/3010) Feature: Docker Java Debug Option
