+++ security/src/test/scala/pl/touk/nussknacker/ui/security/api/LoggedUserTest.scala
@@ -2,12 +2,12 @@ package pl.touk.nussknacker.ui.security.api
 
 import org.scalatest.{FunSuite, Matchers}
 import org.scalatest.prop.TableDrivenPropertyChecks._
-import Permission._
 import org.scalatest.prop.{TableFor3, TableFor4}
+import pl.touk.nussknacker.security.Permission.Permission
 import pl.touk.nussknacker.ui.security.api.AuthenticationConfiguration.ConfigRule
 
 class LoggedUserTest extends FunSuite with Matchers {
-
+  import pl.touk.nussknacker.security.Permission._
   test("Admin permission grants other permissions") {
     def admin(cp: Map[String, Set[Permission]]) = LoggedUser("1", "admin", cp, Nil, true)
