+++ security/src/test/scala/pl/touk/nussknacker/ui/security/api/RulesSetSpec.scala
@@ -3,8 +3,9 @@ package pl.touk.nussknacker.ui.security.api
 import org.scalatest.prop.TableDrivenPropertyChecks
 import org.scalatest.{FunSuite, Matchers}
 import pl.touk.nussknacker.ui.security.api.AuthenticationConfiguration.ConfigRule
-import pl.touk.nussknacker.ui.security.api.GlobalPermission.GlobalPermission
-import pl.touk.nussknacker.ui.security.api.Permission.Permission
+import GlobalPermission.GlobalPermission
+import pl.touk.nussknacker.security.Permission
+import pl.touk.nussknacker.security.Permission.Permission
 
 
 class RulesSetSpec  extends FunSuite with Matchers with TableDrivenPropertyChecks {
