+++ security/src/test/scala/pl/touk/nussknacker/ui/security/oauth2/DefaultOAuth2ServiceFactorySpec.scala
@@ -1,8 +1,9 @@
 package pl.touk.nussknacker.ui.security.oauth2
 
 import org.scalatest.{FlatSpec, Matchers, Suite}
+import pl.touk.nussknacker.security.Permission
 import pl.touk.nussknacker.test.PatientScalaFutures
-import pl.touk.nussknacker.ui.security.api.{LoggedUser, Permission}
+import pl.touk.nussknacker.ui.security.api.LoggedUser
 import pl.touk.nussknacker.ui.security.oauth2.OAuth2ErrorHandler.{OAuth2CompoundException, OAuth2ServerError}
 import sttp.client.Response
 import sttp.client.testing.SttpBackendStub
