+++ ui/server/src/main/scala/pl/touk/nussknacker/ui/api/AuthorizeProcessDirectives.scala
@@ -3,8 +3,9 @@ package pl.touk.nussknacker.ui.api
 import akka.http.scaladsl.server.{Directive0, Directives}
 import pl.touk.nussknacker.restmodel.process.ProcessIdWithName
 import pl.touk.nussknacker.engine.api.process.ProcessId
-import pl.touk.nussknacker.ui.security.api.Permission.Permission
-import pl.touk.nussknacker.ui.security.api.{LoggedUser, Permission}
+import pl.touk.nussknacker.security.Permission
+import pl.touk.nussknacker.security.Permission.Permission
+import pl.touk.nussknacker.ui.security.api.LoggedUser
 
 import scala.language.implicitConversions
